% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl.R
\name{dl}
\alias{dl}
\title{Download 4 files for FRB_CESAB workshop}
\usage{
dl(url1, url2, url3, url4)
}
\arguments{
\item{url1}{Url of the pantheria file.}

\item{url2}{Url of the wildfinder ecoregions list.}

\item{url3}{Url of the wildfinder ecoregions species.}

\item{url4}{Url of the wildfinder ecoregions species.}
}
\description{
This function downloads the PanTHERIA dataset hosted on the
GitHub repository \url{https://github.com/rdatatoolbox/datarepo/}.
}
